backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-dotenv
langchain
langchain-openai
openai
pydantic
